+++ src/test/java/com/sparta/delivery/domain/menu/service/MenuServiceTest.java
@@ -1,6 +1,6 @@
 package com.sparta.delivery.domain.menu.service;
 
-import com.sparta.delivery.common.SignUser;
+import com.sparta.delivery.etc.common.SignUser;
 import com.sparta.delivery.menu.dto.*;
 import com.sparta.delivery.menu.entity.Menu;
 import com.sparta.delivery.menu.enums.MenuStatus;
